Move common parameters to datatypes
